backend/requirements.txt
# Python Backend for UniversalChatbot

# Core Dependencies
openai>=1.0.0
python-dotenv>=0.19.0
pyyaml>=6.0

# Development
pytest>=7.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0
black>=22.0
flake8>=5.0

# Logging
python-json-logger>=2.0.0